vocabDrillページをSSR化
